# Swagger/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Api de Tanks de Guerra",
    "version": "1.0.0",
    "description": "API para cadastro de Tanks de Guerra"
  },
  "paths": {
    "/tanks": {
      "get": {
        "tags": ["Operações sem ID"],
        "summary": "Obter a lista de Tanks",
        "responses": {
          "202": {
            "description": "Lista de Tanks retornada com sucesso"
          }
        }
      },
      "post": {
        "tags": ["Operações sem ID"],
        "summary": "Criar um novo Tank",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/tanks"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Novo Tank cadastrado com sucesso!"
          },
          "400": {
            "description": "ID já existe, tente novamente com um ID diferente"
          }
        }
      }
    },
    "/tanks/{id}": {
      "put": {
        "tags": ["Operações com ID"],
        "summary": "Atualizar informações de um Tank pelo ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/tanks"
              }
            }
          }
        },
        "responses": {
          "404": {
            "description": "Tank não encontrada :/"
          }
        }
      },
      "delete": {
        "tags": ["Operações com ID"],
        "summary": "Excluir um Tank pelo ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Tank excluído com sucesso"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Tank": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "titulo": {
            "type": "string"
          },
          "descricao": {
            "type": "string"
          },
          "exemplo": {
            "type": "string"
          },
          "autor": {
            "type": "string"
          },
          "ano": {
            "type": "string"
          },
          "pais": {
            "type": "string"
          }
        }
      }
    }
  }
}